Int型からDouble型に変更

+++ 開発project/ViewController.swift
@@ -17,11 +17,11 @@ class ViewController: UIViewController {
     @IBOutlet private var resultLabel: UILabel!
     
     @IBAction func calcButton(_ sender: Any) {
-        let number1 = Int(textField1.text!) ?? 0
-        let number2 = Int(textField2.text!) ?? 0
-        let number3 = Int(textField3.text!) ?? 0
-        let number4 = Int(textField4.text!) ?? 0
-        let number5 = Int(textField5.text!) ?? 0
+        let number1 = Double(textField1.text!) ?? 0
+        let number2 = Double(textField2.text!) ?? 0
+        let number3 = Double(textField3.text!) ?? 0
+        let number4 = Double(textField4.text!) ?? 0
+        let number5 = Double(textField5.text!) ?? 0
         let result = number1 + number2 + number3 + number4 + number5
         resultLabel.text = String(result)
